+++ src/commands/mod.rs
@@ -14,3 +14,4 @@ pub mod pull;
 pub mod run;
 pub mod version;
 pub mod world;
+pub mod ws;