+++ src/commands/ws/mod.rs
@@ -1,0 +1,9 @@
+use anyhow::Result;
+
+use crate::app::App;
+
+pub async fn run(mut app: App) -> Result<()> {
+    
+    
+    Ok(())
+}