+++ src/interop/mod.rs
@@ -3,3 +3,4 @@ pub mod markdown;
 pub mod mrpack;
 pub mod packwiz;
 pub mod worlds;
+pub mod hooks;